Refactor: Organize project structure
