+++ composeApp/src/commonMain/kotlin/org/example/project/App.kt
@@ -2,7 +2,7 @@ package org.example.project
 
 import androidx.compose.runtime.Composable
 import org.example.project.presentation.designsystem.textstyle.AppTheme
-import org.example.project.presentation.ui.screens.categoryScreen.AccountSetupCategoryScreen
+import org.example.project.presentation.ui.screens.setupScreens.AccountSetupCategoryScreen
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import org.koin.compose.KoinContext
