+++ composeApp/src/commonMain/kotlin/org/example/project/data/repository/dataSource/CategoryDataSource.kt
@@ -1,6 +1,6 @@
 package org.example.project.data.repository.dataSource
 
-import org.example.project.data.repository.dataSource.local.dto.CategoryEntity
+import org.example.project.data.repository.dataSource.memory.dto.CategoryEntity
 
 interface CategoryDataSource {
     suspend fun getCategories(): List<CategoryEntity>
