+++ composeApp/src/commonMain/kotlin/org/example/project/data/repository/dataSource/memory/dto/CategoryEntity.kt
@@ -1,4 +1,4 @@
-package org.example.project.data.repository.dataSource.local.dto
+package org.example.project.data.repository.dataSource.memory.dto
 
 import androidx.compose.ui.graphics.Color
