+++ composeApp/src/commonMain/kotlin/org/example/project/data/repository/mapper/Category.kt
@@ -1,6 +1,6 @@
 package org.example.project.data.repository.mapper
 
-import org.example.project.data.repository.dataSource.local.dto.CategoryEntity
+import org.example.project.data.repository.dataSource.memory.dto.CategoryEntity
 import org.example.project.domain.entity.Category
 
 fun Category.toCategoryEntity(): CategoryEntity {
